[Feature] Search 탭 관련 화면들의 DIContainer 적용

+++ travelPlan/Source/Data/Repositories/SpringServer/Search/Protocol/DestinationScrapRepository.swift
@@ -12,8 +12,8 @@ protocol DestinationScrapRepository {
   /// 폴더이름에 해당하는 스크랩된 여행지 리스트를 가져옵니다.
   func getDestinationScrapList(
     folderName: String,
-    page: Int32?,
-    perPage: Int32?
+    page: Int?,
+    perPage: Int?
   ) -> AnyPublisher<[DestinationScrapDetail], any Error>
   
   /// 여행지 스크랩 추가/삭제를 요청합니다.
@@ -21,7 +21,7 @@ protocol DestinationScrapRepository {
   /// 스크랩 추가 시, folderName이 필수로 요구됩니다.
   /// 반면 스크랩 삭제 시, folderName을 요구하지 않습니다.
   func toggleDestinationScrap(
-    id: Int64,
+    id: Int,
     folderName: String?
   ) -> AnyPublisher<DestinationScrapToggler, any Error>
   
@@ -29,7 +29,7 @@ protocol DestinationScrapRepository {
   ///
   /// folderName으로 objectIdList를 옮깁니다.
   func updateDestinationScrap(
-    objectIdList: [Int64],
+    objectIdList: [Int],
     folderName: String
   ) -> AnyPublisher<[UpdatedDestinationScrap], any Error>
 }
