+++ travelPlan/Source/Domain/UseCases/Search/DefaultDestinationSearchResultUseCase.swift
@@ -26,14 +26,14 @@ final class DefaultDestinationSearchResultUseCase {
 extension DefaultDestinationSearchResultUseCase: DestinationSearchResultUseCase {
   func fetchDestinationList(
     keyword: String,
-    page: Int32?,
-    perPage: Int32?
+    page: Int?,
+    perPage: Int?
   ) -> AnyPublisher<[ThumbnailDestination], any Error> {
     return searchRepository.fetchDestinationList(by: keyword, page: page, perPage: perPage)
   }
   
   func toggleScrap(
-    id: Int64,
+    id: Int,
     folderName: String?
   ) -> AnyPublisher<DestinationScrapToggler, any Error> {
     return scrapRepository.toggleDestinationScrap(id: id, folderName: folderName)
